requirements.txt
flask
flask-cors
tensorflow
numpy
pandas
Pillow